adding some drivers

+++ MCAL_layer/mem_map.h
@@ -16,7 +16,7 @@
 /*---------------------------------------Data types-------------------------------------------*/
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned int uint32;
+typedef unsigned long int uint32;
 typedef signed char sint8;
 typedef signed short sint16;
 typedef signed int sint32;
